Always copy translation map templates when running in development and test

+++ spec/services/location_maps_generator_service_spec.rb
@@ -5,10 +5,12 @@ RSpec.describe LocationMapsGeneratorService do
   let(:base_path) { './tmp/' }
   let(:locations_path) { File.join(base_path, 'locations.rb') }
   let(:location_display_path) { File.join(base_path, 'location_display.rb') }
+  let(:holding_library_path) { File.join(base_path, 'holding_library.rb') }
 
   after do
-    FileUtils.rm(locations_path)
-    FileUtils.rm(location_display_path)
+    FileUtils.rm(locations_path) if File.exist?(locations_path)
+    FileUtils.rm(location_display_path) if File.exist?(location_display_path)
+    FileUtils.rm(holding_library_path) if File.exist?(holding_library_path)
   end
 
   describe '.generate' do
@@ -45,6 +47,21 @@ RSpec.describe LocationMapsGeneratorService do
       described_class.generate_from_templates
       expect(File.exist?(locations_path)).to be true
       expect(File.exist?(location_display_path)).to be true
+      expect(File.exist?(holding_library_path)).to be true
+    end
+
+    context 'with existing translation map files' do
+      before do
+        allow(FileUtils).to receive(:cp)
+      end
+
+      it 'overwrites the translation map' do
+        # Create a translation map file
+        File.open(locations_path, 'w')
+
+        described_class.generate_from_templates
+        expect(FileUtils).to have_received(:cp).exactly(3).times
+      end
     end
   end
 end
